chore: remove blueprint handling from UI code
Blueprints have been disabled for quite some time. Right now, they are just a nuisance for UI code.
They will need be re-thought and re-built either way, so I'm just removing them from being possible
to be used in a node here.
