+++ web_src/src/lib/layout/types.ts
@@ -1,9 +1,4 @@
-import type {
-  BlueprintsBlueprint,
-  CanvasesCanvas,
-  SuperplaneActionsAction,
-  SuperplaneComponentsNode,
-} from "@/api-client";
+import type { CanvasesCanvas, SuperplaneActionsAction, SuperplaneComponentsNode } from "@/api-client";
 
 export type LayoutScope = "full-canvas" | "connected-component";
 
@@ -11,7 +6,6 @@ export type LayoutEngineApplyOptions = {
   nodeIds?: string[];
   scope?: LayoutScope;
   components?: SuperplaneActionsAction[];
-  blueprints?: BlueprintsBlueprint[];
 };
 
 export interface LayoutEngine {
