+++ web_src/src/pages/workflowv2/index.tsx
@@ -13,7 +13,6 @@ import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { Link, useNavigate, useParams, useSearchParams } from "react-router-dom";
 
 import type {
-  BlueprintsBlueprint,
   CanvasChangesetChange,
   CanvasesCanvas,
   CanvasesCanvasChangeRequest,
@@ -35,7 +34,7 @@ import { useOrganizationRoles, useOrganizationUsers } from "@/hooks/useOrganizat
 
 import { Button } from "@/components/ui/button";
 import { usePermissions } from "@/contexts/PermissionsContext";
-import { useBlueprints, useComponents } from "@/hooks/useBlueprintData";
+import { useComponents } from "@/hooks/useComponentData";
 import {
   canvasKeys,
   eventExecutionsQueryOptions,
@@ -89,13 +88,7 @@ import { getChangeRequestReviewPhase } from "./changeRequestReviewActions";
 import { buildDraftNodeDiffSummary, hasDraftVersusLiveGraphDiff } from "./draftNodeDiff";
 import { prepareAnnotationNode } from "./lib/canvas-annotation-node";
 import { shouldPreserveDraftSpec } from "./lib/draft-canvas-sync";
-import {
-  CANVAS_BUNDLE_COLOR,
-  CANVAS_BUNDLE_ICON_SLUG,
-  prepareComponentNode,
-  prepareCompositeNode,
-  prepareTriggerNode,
-} from "./lib/canvas-node-preparation";
+import { prepareComponentNode, prepareTriggerNode } from "./lib/canvas-node-preparation";
 import {
   isDraftVersion,
   isPublishedVersion,
@@ -191,7 +184,6 @@ export function WorkflowPageV2() {
   const actOnCanvasChangeRequestMutation = useActOnCanvasChangeRequest(organizationId!, canvasId!);
   const resolveCanvasChangeRequestMutation = useResolveCanvasChangeRequest(organizationId!, canvasId!);
   const { data: triggers = [], isLoading: triggersLoading } = useTriggers();
-  const { data: blueprints = [], isLoading: blueprintsLoading } = useBlueprints(organizationId!);
   const { data: components = [], isLoading: componentsLoading } = useComponents(organizationId!);
   const { data: widgets = [], isLoading: widgetsLoading } = useWidgets();
   const { data: availableIntegrations = [], isLoading: integrationsLoading } = useAvailableIntegrations();
@@ -1449,10 +1441,9 @@ export function WorkflowPageV2() {
         scope: "connected-component",
         nodeIds: [nodeID],
         components,
-        blueprints,
       });
     },
-    [isAutoLayoutOnUpdateEnabled, components, blueprints],
+    [isAutoLayoutOnUpdateEnabled, components],
   );
 
   /**
@@ -1698,14 +1689,13 @@ export function WorkflowPageV2() {
   const canvasMode = hasEditableVersion ? "edit" : "live";
 
   const { nodes: preparedNodes, edges } = useMemo(() => {
-    if (!canvas || canvasLoading || triggersLoading || blueprintsLoading || componentsLoading || integrationsLoading) {
+    if (!canvas || canvasLoading || triggersLoading || componentsLoading || integrationsLoading) {
       return { nodes: [], edges: [] };
     }
 
     return prepareData(
       canvas,
       allTriggers,
-      blueprints,
       allComponents,
       visibleNodeEventsMap,
       visibleNodeExecutionsMap,
@@ -1718,7 +1708,6 @@ export function WorkflowPageV2() {
   }, [
     canvas,
     allTriggers,
-    blueprints,
     allComponents,
     visibleNodeEventsMap,
     visibleNodeExecutionsMap,
@@ -1727,7 +1716,6 @@ export function WorkflowPageV2() {
     queryClient,
     canvasLoading,
     triggersLoading,
-    blueprintsLoading,
     componentsLoading,
     integrationsLoading,
     organizationId,
@@ -1765,7 +1753,6 @@ export function WorkflowPageV2() {
       const sidebarData = prepareSidebarData(
         node,
         canvas?.spec?.nodes || [],
-        blueprints,
         allComponents,
         allTriggers,
         executionsMap,
@@ -1781,7 +1768,7 @@ export function WorkflowPageV2() {
         isLoading: nodeData.isLoading,
       };
     },
-    [canvas, blueprints, allComponents, allTriggers, visibleNodeEventsMap, isViewingLiveVersion, getNodeData],
+    [canvas, allComponents, allTriggers, visibleNodeEventsMap, isViewingLiveVersion, getNodeData],
   );
 
   // Trigger data loading when sidebar opens for a node
@@ -2550,11 +2537,7 @@ export function WorkflowPageV2() {
       let integrationLabel: string | undefined;
       let blockName: string | undefined;
 
-      if (node.type === "TYPE_BLUEPRINT") {
-        const blueprintMetadata = blueprints.find((b) => b.id === node.component);
-        configurationFields = blueprintMetadata?.configuration || [];
-        displayLabel = blueprintMetadata?.name || displayLabel;
-      } else if (node.type === "TYPE_ACTION") {
+      if (node.type === "TYPE_ACTION") {
         const componentMetadata = allComponents.find((c) => c.name === node.component);
         configurationFields = componentMetadata?.configuration || [];
         displayLabel = componentMetadata?.label || displayLabel;
@@ -2607,7 +2590,7 @@ export function WorkflowPageV2() {
         integrationRef: node.integration,
       };
     },
-    [canvas, blueprints, allComponents, allTriggers, availableIntegrations, widgets],
+    [canvas, allComponents, allTriggers, availableIntegrations, widgets],
   );
 
   const createIntegrationMutation = useCreateIntegration(organizationId ?? "");
@@ -3072,11 +3055,9 @@ export function WorkflowPageV2() {
         type:
           buildingBlock.type === "trigger"
             ? "TYPE_TRIGGER"
-            : buildingBlock.type === "blueprint"
-              ? "TYPE_BLUEPRINT"
-              : buildingBlock.name === "annotation"
-                ? "TYPE_WIDGET"
-                : "TYPE_ACTION",
+            : buildingBlock.name === "annotation"
+              ? "TYPE_WIDGET"
+              : "TYPE_ACTION",
         configuration: filteredConfiguration,
         integration: integrationRef,
         position: position
@@ -3099,8 +3080,6 @@ export function WorkflowPageV2() {
         newNode.component = buildingBlock.name;
       } else if (buildingBlock.type === "trigger") {
         newNode.component = buildingBlock.name;
-      } else if (buildingBlock.type === "blueprint") {
-        newNode.component = buildingBlock.id;
       }
 
       // Add the new node to the workflow
@@ -3284,7 +3263,7 @@ export function WorkflowPageV2() {
         id: newNodeId,
         name: placeholderName,
         type: "TYPE_ACTION",
-        // NO component/blueprint/trigger reference - causes validation error
+        // NO component/trigger reference - causes validation error
         configuration: {},
         metadata: {},
         position: {
@@ -3365,12 +3344,7 @@ export function WorkflowPageV2() {
       const updatedNode: ComponentsNode = {
         ...canvas.spec!.nodes![nodeIndex],
         name: uniqueNodeName,
-        type:
-          data.buildingBlock.type === "trigger"
-            ? "TYPE_TRIGGER"
-            : data.buildingBlock.type === "blueprint"
-              ? "TYPE_BLUEPRINT"
-              : "TYPE_ACTION",
+        type: data.buildingBlock.type === "trigger" ? "TYPE_TRIGGER" : "TYPE_ACTION",
         configuration: filteredConfiguration,
       };
 
@@ -3379,8 +3353,6 @@ export function WorkflowPageV2() {
         updatedNode.component = data.buildingBlock.name;
       } else if (data.buildingBlock.type === "trigger") {
         updatedNode.component = data.buildingBlock.name;
-      } else if (data.buildingBlock.type === "blueprint") {
-        updatedNode.component = data.buildingBlock.id;
       }
 
       const updatedNodes = [...(canvas.spec?.nodes || [])];
@@ -3539,7 +3511,6 @@ export function WorkflowPageV2() {
         nodeIds,
         scope: "connected-component",
         components,
-        blueprints,
       });
 
       applyLocalWorkflowUpdate(updatedWorkflow);
@@ -3548,16 +3519,7 @@ export function WorkflowPageV2() {
         await handleSaveWorkflow(updatedWorkflow, { showToast: false });
       }
     },
-    [
-      canvas,
-      components,
-      blueprints,
-      organizationId,
-      canvasId,
-      handleSaveWorkflow,
-      isReadOnly,
-      applyLocalWorkflowUpdate,
-    ],
+    [canvas, components, organizationId, canvasId, handleSaveWorkflow, isReadOnly, applyLocalWorkflowUpdate],
   );
 
   const handleNodesDuplicate = useCallback(
@@ -3580,9 +3542,6 @@ export function WorkflowPageV2() {
             baseName = nodeToDuplicate.component;
           } else if (nodeToDuplicate.type === "TYPE_ACTION" && nodeToDuplicate.component) {
             baseName = nodeToDuplicate.component;
-          } else if (nodeToDuplicate.type === "TYPE_BLUEPRINT" && nodeToDuplicate.component) {
-            const blueprintMetadata = blueprints.find((b) => b.id === nodeToDuplicate.component);
-            baseName = blueprintMetadata?.name || "blueprint";
           } else {
             baseName = "node";
           }
@@ -3638,7 +3597,7 @@ export function WorkflowPageV2() {
         await handleSaveWorkflow(updatedWorkflow, { showToast: false });
       }
     },
-    [canvas, organizationId, canvasId, blueprints, handleSaveWorkflow, isReadOnly, applyLocalWorkflowUpdate],
+    [canvas, organizationId, canvasId, handleSaveWorkflow, isReadOnly, applyLocalWorkflowUpdate],
   );
 
   const handleEdgeDelete = useCallback(
@@ -3923,10 +3882,6 @@ export function WorkflowPageV2() {
           baseName = nodeToDuplicate.component;
         } else if (nodeToDuplicate.type === "TYPE_ACTION" && nodeToDuplicate.component) {
           baseName = nodeToDuplicate.component;
-        } else if (nodeToDuplicate.type === "TYPE_BLUEPRINT" && nodeToDuplicate.component) {
-          // For blueprints, we need to find the blueprint metadata to get the name
-          const blueprintMetadata = blueprints.find((b) => b.id === nodeToDuplicate.component);
-          baseName = blueprintMetadata?.name || "blueprint";
         } else {
           baseName = "node";
         }
@@ -3975,7 +3930,6 @@ export function WorkflowPageV2() {
       canvas,
       organizationId,
       canvasId,
-      blueprints,
       handleSaveWorkflow,
       applyAutoLayoutOnAddedNode,
       isReadOnly,
@@ -4883,8 +4837,6 @@ export function WorkflowPageV2() {
         componentName = node.component;
       } else if (node.type === "TYPE_TRIGGER" && node.component) {
         componentName = node.component;
-      } else if (node.type === "TYPE_BLUEPRINT" && node.component) {
-        componentName = "default";
       }
 
       return {
@@ -4905,8 +4857,6 @@ export function WorkflowPageV2() {
         componentName = node.component;
       } else if (node.type === "TYPE_ACTION" && node.component) {
         componentName = node.component;
-      } else if (node.type === "TYPE_BLUEPRINT" && node.component) {
-        componentName = "default";
       }
 
       const renderer = getCustomFieldRenderer(componentName);
@@ -4993,8 +4943,7 @@ export function WorkflowPageV2() {
   );
 
   const isInitialCanvasBootstrapLoading =
-    !canvas &&
-    (canvasLoading || triggersLoading || blueprintsLoading || componentsLoading || widgetsLoading || usersLoading);
+    !canvas && (canvasLoading || triggersLoading || componentsLoading || widgetsLoading || usersLoading);
   const isDraftCanvasLoading =
     isViewingDraftVersion &&
     !!activeCanvasVersionId &&
@@ -5345,7 +5294,6 @@ export function WorkflowPageV2() {
           workflowNodes={canvas?.spec?.nodes}
           components={allComponents}
           triggers={allTriggers}
-          blueprints={blueprints}
           logEntries={logEntries}
           runsEvents={isViewingLiveVersion ? runsEventsData.events : []}
           runsTotalCount={isViewingLiveVersion ? runsEventsData.totalCount : 0}
@@ -5616,7 +5564,6 @@ function getNodesBeforeTarget(targetNodeId: string, workflow: CanvasesCanvas): S
 function prepareData(
   workflow: CanvasesCanvas,
   triggers: TriggersTrigger[],
-  blueprints: BlueprintsBlueprint[],
   components: SuperplaneActionsAction[],
   nodeEventsMap: Record<string, CanvasesCanvasEvent[]>,
   nodeExecutionsMap: Record<string, CanvasesCanvasNodeExecution[]>,
@@ -5640,7 +5587,6 @@ function prepareData(
           workflowNodes,
           node,
           triggers,
-          blueprints,
           components,
           nodeEventsMap,
           nodeExecutionsMap,
@@ -5664,7 +5610,6 @@ function prepareNode(
   nodes: ComponentsNode[],
   node: ComponentsNode,
   triggers: TriggersTrigger[],
-  blueprints: BlueprintsBlueprint[],
   components: SuperplaneActionsAction[],
   nodeEventsMap: Record<string, CanvasesCanvasEvent[]>,
   nodeExecutionsMap: Record<string, CanvasesCanvasNodeExecution[]>,
@@ -5678,23 +5623,6 @@ function prepareNode(
   switch (node.type) {
     case "TYPE_TRIGGER":
       return prepareTriggerNode(node, triggers, nodeEventsMap, canvasMode);
-    case "TYPE_BLUEPRINT": {
-      const componentMetadata = components.find((c) => c.name === node.component);
-      const compositeNode = prepareCompositeNode(nodes, node, blueprints, nodeExecutionsMap, nodeQueueItemsMap);
-
-      // Override outputChannels with component metadata if available
-      if (componentMetadata?.outputChannels) {
-        return {
-          ...compositeNode,
-          data: {
-            ...compositeNode.data,
-            outputChannels: componentMetadata.outputChannels.map((c) => c.name!),
-          },
-        };
-      }
-
-      return compositeNode;
-    }
     case "TYPE_WIDGET":
       return prepareAnnotationNode(node);
 
@@ -5728,7 +5656,6 @@ function prepareEdge(edge: ComponentsEdge): CanvasEdge {
 function prepareSidebarData(
   node: ComponentsNode,
   nodes: ComponentsNode[],
-  blueprints: BlueprintsBlueprint[],
   components: SuperplaneActionsAction[],
   triggers: TriggersTrigger[],
   nodeExecutionsMap: Record<string, CanvasesCanvasNodeExecution[]>,
@@ -5742,20 +5669,14 @@ function prepareSidebarData(
   const events = nodeEventsMap[node.id!] || [];
 
   // Get metadata based on node type
-  const blueprintMetadata =
-    node.type === "TYPE_BLUEPRINT" ? blueprints.find((b) => b.id === node.component) : undefined;
   const componentMetadata = node.type === "TYPE_ACTION" ? components.find((c) => c.name === node.component) : undefined;
   const triggerMetadata = node.type === "TYPE_TRIGGER" ? triggers.find((t) => t.name === node.component) : undefined;
 
-  const nodeTitle =
-    componentMetadata?.label || blueprintMetadata?.name || triggerMetadata?.label || node.name || "Unknown";
+  const nodeTitle = componentMetadata?.label || triggerMetadata?.label || node.name || "Unknown";
   let iconSlug = "boxes";
   let color = "indigo";
 
-  if (blueprintMetadata) {
-    iconSlug = CANVAS_BUNDLE_ICON_SLUG;
-    color = CANVAS_BUNDLE_COLOR;
-  } else if (componentMetadata) {
+  if (componentMetadata) {
     iconSlug = componentMetadata.icon || iconSlug;
     color = componentMetadata.color || color;
   } else if (triggerMetadata) {
@@ -5781,6 +5702,6 @@ function prepareSidebarData(
     totalInHistoryCount: totalHistoryCount ? totalHistoryCount : 0,
     totalInQueueCount: totalQueueCount ? totalQueueCount : 0,
     hideQueueEvents,
-    isComposite: node.type === "TYPE_BLUEPRINT",
+    isComposite: false,
   };
 }
