+++ web_src/src/pages/workflowv2/lib/canvas-node-preparation.ts
@@ -2,7 +2,6 @@ import type { QueryClient } from "@tanstack/react-query";
 import { Puzzle } from "lucide-react";
 import {
   canvasesInvokeNodeExecutionHook,
-  type BlueprintsBlueprint,
   type CanvasesCanvasEvent,
   type CanvasesCanvasNodeExecution,
   type CanvasesCanvasNodeQueueItem,
@@ -16,13 +15,10 @@ import { getApiErrorMessage } from "@/lib/errors";
 import { showErrorToast } from "@/lib/toast";
 import { getHeaderIconSrc } from "@/ui/componentSidebar/integrationIcons";
 import type { CanvasNode } from "@/ui/CanvasPage";
-import type { CompositeProps, LastRunState } from "@/ui/composite";
 import type { ActionContext, ComponentBaseMapper, User } from "../mappers/types";
 import { getComponentBaseMapper, getTriggerRenderer } from "../mappers";
 import { buildComponentFallbackCanvasNode, buildTriggerFallbackCanvasNode } from "./canvas-node-fallback";
 
-export const CANVAS_BUNDLE_ICON_SLUG = "component";
-export const CANVAS_BUNDLE_COLOR = "gray";
 import {
   buildComponentDefinition,
   buildEventInfo,
@@ -30,7 +26,6 @@ import {
   buildNodeInfo,
   buildQueueItemInfo,
   buildUserInfo,
-  getNextInQueueInfo,
 } from "../utils";
 import { canvasKeys } from "@/hooks/useCanvasData";
 import { withOrganizationHeader } from "@/lib/withOrganizationHeader";
@@ -66,18 +61,6 @@ type NodePosition = {
   x: number;
   y: number;
 };
-
-function getRunItemState(execution: CanvasesCanvasNodeExecution): LastRunState {
-  if (execution.state === "STATE_PENDING" || execution.state === "STATE_STARTED") {
-    return "running";
-  }
-
-  if (execution.state === "STATE_FINISHED" && execution.result === "RESULT_PASSED") {
-    return "success";
-  }
-
-  return "failed";
-}
 
 function getNodePosition(node: ComponentsNode): NodePosition {
   return {
@@ -124,110 +107,6 @@ function buildPreparedTriggerCanvasNode(args: {
       },
     },
   };
-}
-
-function buildCompositeFieldLabelMap(blueprintMetadata?: BlueprintsBlueprint): Record<string, string> {
-  const configurationFields = blueprintMetadata?.configuration || [];
-
-  return configurationFields.reduce<Record<string, string>>((acc, field) => {
-    if (field.name) {
-      acc[field.name] = field.label || field.name;
-    }
-    return acc;
-  }, {});
-}
-
-function buildCompositeParameters(
-  node: ComponentsNode,
-  fieldLabelMap: Record<string, string>,
-): CompositeProps["parameters"] {
-  const configuration = node.configuration || {};
-  const configurationKeys = Object.keys(configuration);
-
-  if (configurationKeys.length === 0) {
-    return [];
-  }
-
-  return [
-    {
-      icon: "cog",
-      items: configurationKeys.reduce(
-        (acc, key) => {
-          const displayKey = fieldLabelMap[key] || key;
-          acc[displayKey] = `${configuration[key]}`;
-          return acc;
-        },
-        {} as Record<string, string>,
-      ),
-    },
-  ];
-}
-
-function buildCompositeCanvasNode(args: {
-  node: ComponentsNode;
-  displayLabel: string;
-  blueprintMetadata?: BlueprintsBlueprint;
-  isMissing: boolean;
-  position: { x: number; y: number };
-  parameters: CompositeProps["parameters"];
-}): CanvasNode {
-  const { node, displayLabel, blueprintMetadata, isMissing, position, parameters } = args;
-
-  return {
-    id: node.id!,
-    position,
-    data: {
-      type: "composite",
-      label: displayLabel,
-      state: "pending" as const,
-      outputChannels: blueprintMetadata?.outputChannels?.map((c) => c.name!) || ["default"],
-      composite: {
-        iconSlug: CANVAS_BUNDLE_ICON_SLUG,
-        iconColor: getColorClass(CANVAS_BUNDLE_COLOR),
-        collapsedBackground: getBackgroundColorClass(CANVAS_BUNDLE_COLOR),
-        collapsed: node.isCollapsed,
-        title: displayLabel,
-        description: blueprintMetadata?.description,
-        isMissing,
-        error: node.errorMessage,
-        warning: node.warningMessage,
-        paused: !!node.paused,
-        parameters,
-      },
-    },
-  };
-}
-
-function appendCompositeLastRunItem(
-  canvasNode: CanvasNode,
-  execution: CanvasesCanvasNodeExecution,
-  nodes: ComponentsNode[],
-) {
-  const rootTriggerNode = nodes.find((n) => n.id === execution.rootEvent?.nodeId);
-  const rootTriggerRenderer = getTriggerRenderer(rootTriggerNode?.component || "");
-  const eventInfo = buildEventInfo(execution.rootEvent!);
-  const { title, subtitle } = rootTriggerRenderer.getTitleAndSubtitle({ event: eventInfo });
-
-  (canvasNode.data.composite as CompositeProps).lastRunItem = {
-    title,
-    subtitle,
-    id: execution.rootEvent?.id,
-    receivedAt: new Date(execution.createdAt!),
-    state: getRunItemState(execution),
-    values: rootTriggerRenderer.getRootEventValues({ event: eventInfo }),
-  };
-}
-
-function appendCompositeQueueInfo(
-  canvasNode: CanvasNode,
-  nodeId: string,
-  nodeQueueItemsMap: Record<string, CanvasesCanvasNodeQueueItem[]>,
-  nodes: ComponentsNode[],
-) {
-  const nextInQueueInfo = getNextInQueueInfo(nodeQueueItemsMap, nodeId, nodes);
-  if (nextInQueueInfo) {
-    (canvasNode.data.composite as CompositeProps).nextInQueue = nextInQueueInfo;
-  }
 }
 
 function buildPlaceholderComponentNode(node: ComponentsNode): CanvasNode {
@@ -300,39 +179,6 @@ export function prepareTriggerNode(
     console.error(`[CanvasPage] Failed to prepare trigger node "${node.id}":`, error);
     return buildTriggerFallbackCanvasNode({ node, displayLabel, triggerMetadata });
   }
-}
-
-export function prepareCompositeNode(
-  nodes: ComponentsNode[],
-  node: ComponentsNode,
-  blueprints: BlueprintsBlueprint[],
-  nodeExecutionsMap: Record<string, CanvasesCanvasNodeExecution[]>,
-  nodeQueueItemsMap: Record<string, CanvasesCanvasNodeQueueItem[]>,
-): CanvasNode {
-  const blueprintMetadata = blueprints.find((b) => b.id === node.component);
-  const isMissing = !blueprintMetadata;
-  const executions = nodeExecutionsMap[node.id!] || [];
-  const displayLabel = node.name || blueprintMetadata?.name || "Composite";
-  const position = getNodePosition(node);
-
-  const fieldLabelMap = buildCompositeFieldLabelMap(blueprintMetadata);
-  const parameters = buildCompositeParameters(node, fieldLabelMap);
-  const canvasNode = buildCompositeCanvasNode({
-    node,
-    displayLabel,
-    blueprintMetadata,
-    isMissing,
-    position,
-    parameters,
-  });
-
-  if (executions.length > 0) {
-    appendCompositeLastRunItem(canvasNode, executions[0], nodes);
-  }
-
-  appendCompositeQueueInfo(canvasNode, node.id!, nodeQueueItemsMap, nodes);
-
-  return canvasNode;
 }
 
 export function prepareComponentNode(args: PrepareComponentNodeArgs): CanvasNode {
