+++ web_src/src/ui/BuildingBlocksSidebar/types.ts
@@ -1,11 +1,11 @@
-import type { SuperplaneBlueprintsOutputChannel, SuperplaneActionsOutputChannel } from "@/api-client";
+import type { SuperplaneActionsOutputChannel } from "@/api-client";
 
 export interface BuildingBlock {
   name: string;
   label?: string;
   description?: string;
-  type: "trigger" | "component" | "blueprint";
-  outputChannels?: Array<SuperplaneActionsOutputChannel | SuperplaneBlueprintsOutputChannel>;
+  type: "trigger" | "component";
+  outputChannels?: Array<SuperplaneActionsOutputChannel>;
   configuration?: any[];
   icon?: string;
   color?: string;
